requirements.txt
# FastAPI and Server
fastapi==0.128.0
uvicorn[standard]==0.40.0

# Pydantic for data validation
pydantic==2.12.5
pydantic-settings==2.12.0

# Environment variables
python-dotenv==1.2.1

# Google Generative AI (Gemini) - 새로운 패키지
google-genai==1.60.0

# AI Planner V2 (LangGraph & Supabase)
langgraph>=0.2.0
langchain-core>=0.3.0
supabase>=2.0.0
sentence-transformers>=2.2.0
